refactor(value): return from Integer wrapper class int value

+++ stockx/src/main/java/edu/poniperro/stockx/item/Ask.java
@@ -16,7 +16,7 @@ public class Ask implements Offer{
 
     @Override
     public int value() {
-        return this.ask;
+        return this.ask.intValue();
     }
 
     @Override
