+++ stockx/src/main/java/edu/poniperro/stockx/item/Bid.java
@@ -16,7 +16,7 @@ public class Bid implements Offer{
 
     @Override
     public int value() {
-        return this.bid;
+        return this.bid.intValue();
     }
 
     @Override
