+++ stockx/src/main/java/edu/poniperro/stockx/item/Sale.java
@@ -16,7 +16,7 @@ public class Sale implements Offer{
 
     @Override
     public int value() {
-        return this.price;
+        return this.price.intValue();
     }
 
     @Override
